managementgroup: splitting the validate funcs
