+++ azurerm/internal/services/managementgroup/validate/management_group_id.go
@@ -1,0 +1,22 @@
+package validate
+
+import (
+	"fmt"
+
+	"github.com/terraform-providers/terraform-provider-azurerm/azurerm/internal/services/managementgroup/parse"
+)
+
+func ManagementGroupID(i interface{}, k string) (warnings []string, errors []error) {
+	v, ok := i.(string)
+	if !ok {
+		errors = append(errors, fmt.Errorf("expected type of %q to be string", k))
+		return
+	}
+
+	if _, err := parse.ManagementGroupID(v); err != nil {
+		errors = append(errors, fmt.Errorf("Can not parse %q as a management group id: %v", k, err))
+		return
+	}
+
+	return
+}
